+++ app/src/main/java/projekt/substratum/util/readers/ReadCloudShowcaseFile.java
@@ -47,7 +47,7 @@ public enum ReadCloudShowcaseFile {
             final Map<String, String> map = new TreeMap<>();
             for (int temp = 0; temp < nList.getLength(); temp++) {
                 final Node nNode = nList.item(temp);
-                if (nNode.getNodeType() == Node.ELEMENT_NODE) {
+                if ((int) nNode.getNodeType() == (int) Node.ELEMENT_NODE) {
                     final Element eElement = (Element) nNode;
 
                     final String addon_download_name = eElement.getAttribute("id");
