+++ app/src/main/java/projekt/substratum/util/readers/ReadSupportedROMsFile.java
@@ -47,7 +47,7 @@ public enum ReadSupportedROMsFile {
             final Map<String, String> hashMap = new HashMap<>();
             for (int temp = 0; temp < nList.getLength(); temp++) {
                 final Node nNode = nList.item(temp);
-                if (nNode.getNodeType() == Node.ELEMENT_NODE) {
+                if ((int) nNode.getNodeType() == (int) Node.ELEMENT_NODE) {
                     final Element eElement = (Element) nNode;
 
                     final String name = eElement.getAttribute("name");
